+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
 function Header() {
-  return <h1>My React Task List</h1>;
+  return <h1 className="header"> App Mi Lista de Tareas</h1>;
 }
 export default Header;
