+++ src/components/Tareas.jsx
@@ -1,50 +1,19 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import TareasCompletadas from "./TareasCompletadas";
+import { useTareas } from "../hooks/useTareas";
 
 function Tareas() {
-  const [tareas, setTareas] = useState([]);
-  const [tareasCompletadas, setTareasCompletadas] = useState([]);
-  const [tareasLocalStorage, setTareasLocalStorage] = useState(
-    JSON.parse(localStorage.getItem("tareasLocalStorage")) || []
-  );
-
-  useEffect(() => {
-    localStorage.setItem(
-      "tareasLocalStorage",
-      JSON.stringify(tareasLocalStorage)
-    );
-  }, [tareas, tareasCompletadas, tareasLocalStorage]);
-
-  const agregarTarea = (nombre) => {
-    const nuevaTarea = { nombre, completada: false };
-    setTareas([...tareas, nuevaTarea]);
-  };
-
-  const completarTarea = (index) => {
-    const nuevasTareas = [...tareas];
-    nuevasTareas[index].completada = true;
-    const tareaCompletada = nuevasTareas.splice(index, 1)[0];
-    setTareas(nuevasTareas);
-    setTareasCompletadas([...tareasCompletadas, tareaCompletada]);
-  };
-
-  useEffect(() => {
-    localStorage.setItem(
-      "tareasLocalStorage",
-      JSON.stringify({ tareas, tareasCompletadas })
-    );
-  }, [tareas, tareasCompletadas]);
-
-  useEffect(() => {
-    if (tareasLocalStorage) {
-      setTareas(tareasLocalStorage.tareas || []);
-      setTareasCompletadas(tareasLocalStorage.tareasCompletadas || []);
-    }
-  }, [tareasLocalStorage]);
+  const {
+    agregarTarea,
+    completarTarea,
+    eliminarTarea,
+    tareas,
+    tareasCompletadas,
+  } = useTareas();
 
   return (
     <div>
-      <h1>Lista de Tareas</h1>
+      <h2 style={{ color: "green" }}>Detalle de Lista de Tareas:</h2>
       <div>
         <h2>Tareas Pendientes</h2>
         <ul>
@@ -52,6 +21,8 @@ function Tareas() {
             <li key={index}>
               {tarea.nombre}
               <button onClick={() => completarTarea(index)}>Completada</button>
+
+              <button onClick={() => eliminarTarea(index)}>Eliminar</button>
             </li>
           ))}
         </ul>
@@ -63,7 +34,7 @@ function Tareas() {
           }}
         >
           <input name="nuevaTarea" />
-          <button>Agregar Tarea</button>
+          <button>Agregar Nueva Tarea</button>
         </form>
       </div>
       <TareasCompletadas tareasCompletadas={tareasCompletadas} />
