+++ src/components/TareasCompletadas.jsx
@@ -1,19 +1,11 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 function TareasCompletadas({ tareasCompletadas }) {
-  const [tareas, setTareas] = useState([]);
-
-  useEffect(() => {
-    if (tareasCompletadas) {
-      setTareas(tareasCompletadas.filter((tarea) => tarea.completada));
-    }
-  }, [tareasCompletadas]);
-
   return (
     <div>
-      <h2>Tareas Completadas</h2>
+      <h2>Tareas Completadas Exitosamente</h2>
       <ul>
-        {tareas.map((tarea, index) => (
+        {tareasCompletadas.map((tarea, index) => (
           <li key={index}>{tarea.nombre}</li>
         ))}
       </ul>
