+++ src/index.css
@@ -1,6 +1,13 @@
 body {
-  background-color: rgb(216, 220, 246);
+  background-color: rgb(231, 233, 241);
   display: flex;
   justify-content: center;
   margin: 10px 10;
+}
+
+.header {
+  color: rgb(98, 98, 188);
+  font-size: 24px;
+  display: flex;
+  justify-content: center;
 }
